refactor(client):move graph components
