+++ packages/amplication-client/src/Components/GraphComponents/GraphControls.tsx
