+++ packages/amplication-client/src/Components/GraphComponents/GraphToolbar.scss
